requirements.txt
opencv-python
numpy
pillow
# scikit-image  # Optional: for higher-quality dithering, segmentation, etc. 